+++ cuda_src/CCD_CUDA.cu
@@ -280,8 +280,8 @@ cudaError_t ccdpp_NV(smat_t& R_C, testset_t& T, mat_t& W, mat_t& H, parameter& p
         cudaStatus = cudaDeviceSynchronize();
         gpuErrchk(cudaStatus);
 
-        float tot_rmse = 0;
-        float f_rmse = 0;
+        double tot_rmse = 0;
+        double f_rmse = 0;
         gpuErrchk(cudaMemcpy(rmse, d_rmse, (T.nnz + 1) * sizeof(float), cudaMemcpyDeviceToHost));
 
         for (long i = 0; i < T.nnz; ++i) {
